[#1599] Factor Metadata model back to single file
